Refactored package names for clarity of purpose
