+++ src/main/java/com/royware/corona/dashboard/services/chartlistmakers/ChangeInTotalCasesVersusCasesWithExponentialLineChartList.java
@@ -1,4 +1,4 @@
-package com.royware.corona.dashboard.services.chartlists;
+package com.royware.corona.dashboard.services.chartlistmakers;
 
 import java.util.ArrayList;
 import java.util.HashMap;
