+++ src/main/java/com/royware/corona/dashboard/services/chartlistmakers/ChartListStore.java
@@ -1,4 +1,4 @@
-package com.royware.corona.dashboard.services.chartlists;
+package com.royware.corona.dashboard.services.chartlistmakers;
 
 import java.util.List;
 import java.util.Map;
